Move these back into const.py
I'd initially thought I'd be putting a lot more in there than I did
